fix: address Copilot review feedback on PR #69
- resource.rb @example: corrected the inline comment to show the actual
  call (`Adapters::Http::ModelProxy.new(self)`) instead of the
  parenthesised `(self)` syntax that wouldn't paste back as valid Ruby.
- extending.md HTTP bullet: rephrased "no schema methods" to spell out
  that `column_names` is implemented and returns an empty array, so
  callers don't get a misleading impression of the proxy's surface.
