+++ lib/iron_admin.rb
@@ -15,6 +15,7 @@ require "iron_admin/action_field"
 require "iron_admin/nested_association"
 require "iron_admin/nested_attributes_validator"
 require "iron_admin/concerns/nestable"
+require "iron_admin/concerns/soft_deletable"
 require "iron_admin/filters/base_query_builder"
 require "iron_admin/filters/active_record_query_builder"
 require "iron_admin/filters/query_builder"
